+++ packages/ui/Fetch/Fetch.ts
@@ -1,4 +1,4 @@
-import { Base, type BaseConfig, type BaseProps } from '@studiometa/js-toolkit';
+import { Base, type BaseConfig, type BaseProps, BaseInterface } from '@studiometa/js-toolkit';
 import { domScheduler, historyPush, isFunction } from '@studiometa/js-toolkit/utils';
 import morphdom from 'morphdom';
 import { adoptNewScripts, getScripts } from './utils.js';
@@ -14,6 +14,8 @@ export interface FetchProps extends BaseProps {
     headers: Record<string, string>;
     mode: 'replace' | 'prepend' | 'append' | 'morph';
     selector: string;
+    response: string;
+    viewTransition: boolean;
   };
 }
 
@@ -26,7 +28,10 @@ export type FetchConstructor<T extends Fetch = Fetch> = {
  * Fetch class.
  * @link https://ui.studiometa.dev/-/components/Fetch/
  */
-export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps> {
+export class Fetch<T extends BaseProps = BaseProps>
+  extends Base<T & FetchProps>
+  implements BaseInterface
+{
   /**
    * Declare the `this.constructor` type
    * @link https://github.com/microsoft/TypeScript/issues/3841#issuecomment-2381594311
@@ -39,6 +44,7 @@ export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps>
   static FETCH_EVENTS = {
     BEFORE_FETCH: 'fetch-before',
     FETCH: 'fetch-fetch',
+    RESPONSE: 'fetch-response',
     AFTER_FETCH: 'fetch-after',
     BEFORE_UPDATE: 'fetch-update-before',
     UPDATE: 'fetch-update',
@@ -75,6 +81,14 @@ export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps>
       selector: {
         type: String,
         default: '[id]',
+      },
+      response: {
+        type: String,
+        default: 'response.text()',
+      },
+      viewTransition: {
+        type: Boolean,
+        default: true,
       },
     },
   };
@@ -250,7 +264,12 @@ export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps>
     this.__abortController.abort();
     const newController = new AbortController();
     newController.signal.addEventListener('abort', () => {
-      this.$emit(FETCH_EVENTS.ABORT, { instance: this, url, requestInit, reason: newController.signal.reason })
+      this.$emit(FETCH_EVENTS.ABORT, {
+        instance: this,
+        url,
+        requestInit,
+        reason: newController.signal.reason,
+      });
     });
     this.__abortController = newController;
     const init = {
@@ -268,16 +287,24 @@ export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps>
 
     try {
       const response = await this.client(url, init);
+      this.$emit(FETCH_EVENTS.RESPONSE, { instance: this, url, requestInit: init, response });
 
       if (!response.ok) {
         throw new Error(`Fetch failed with status ${response.status}`);
       }
 
-      const content = await response.text();
-      this.$emit(FETCH_EVENTS.AFTER_FETCH, { instance: this, url, requestInit, content });
+      const fn = new Function(
+        'response',
+        'url',
+        'requestInit',
+        'self',
+        `return ${this.$options.response}`,
+      );
+      const content = await fn.call(this, response, url, requestInit, self);
+      this.$emit(FETCH_EVENTS.AFTER_FETCH, { instance: this, url, requestInit: init, content });
       this.update(url, init, content);
     } catch (error) {
-      this.$emit(FETCH_EVENTS.AFTER_FETCH, { instance: this, url, requestInit, error });
+      this.$emit(FETCH_EVENTS.AFTER_FETCH, { instance: this, url, requestInit: init, error });
       this.error(url, init, error);
     }
   }
@@ -327,7 +354,7 @@ export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps>
    */
   async update(url: URL, requestInit: RequestInit, content: string) {
     const { FETCH_EVENTS } = this.constructor;
-    const { history } = this.$options;
+    const { history, viewTransition } = this.$options;
 
     this.$log('content', url, content);
     this.$emit(FETCH_EVENTS.BEFORE_UPDATE, { instance: this, url, requestInit, content });
@@ -347,7 +374,7 @@ export class Fetch<T extends BaseProps = BaseProps> extends Base<T & FetchProps>
 
     this.$emit(FETCH_EVENTS.UPDATE, { instance: this, url, requestInit, fragment });
 
-    if (isFunction(document.startViewTransition)) {
+    if (viewTransition && isFunction(document.startViewTransition)) {
       await document.startViewTransition(() => {
         this.__updateDOM(fragment);
       }).ready;